Index using unsigned ints to eliminate compiler warnings on VS.

+++ src/renderer/thread_pool.cpp.h
@@ -1,10 +1,10 @@
 #include "thread_pool.h"
 
-template<typename parent_class> thread_pool<parent_class>::thread_pool(parent_class *parent, size_t num_workers) :
+template<typename parent_class> thread_pool<parent_class>::thread_pool(parent_class *parent, unsigned num_workers) :
     parent(parent),
     thread_busy(num_workers, false) {
 
-    for (size_t i = 0; i < num_workers; ++i)
+    for (unsigned i = 0; i < num_workers; ++i)
         threads.push_back(thread(&thread_pool::loop, this, i));
 }
 
@@ -20,10 +20,10 @@ template<typename parent_class> thread_pool<parent_class>::~thread_pool() {
     }
 }
 
-template<typename parent_class> void thread_pool<parent_class>::execute(void(parent_class::*func)(size_t thread_index)) {
+template<typename parent_class> void thread_pool<parent_class>::execute(void(parent_class::*func)(unsigned thread_index)) {
     unique_lock<mutex> lock(work_mutex);
     work_function = func;
-    num_threads_busy = threads.size();
+    num_threads_busy = static_cast<unsigned>(threads.size());
 
     for (unsigned i = 0; i < thread_busy.size(); ++i)
         thread_busy[i] = true;
@@ -33,7 +33,7 @@ template<typename parent_class> void thread_pool<parent_class>::execute(void(par
     work_function = nullptr;
 }
 
-template<typename parent_class> void thread_pool<parent_class>::loop(size_t thread_index) {
+template<typename parent_class> void thread_pool<parent_class>::loop(unsigned thread_index) {
     unique_lock<mutex> lock(work_mutex);
 
     while (!stopping) {
