+++ src/renderer/thread_pool.h
@@ -9,13 +9,13 @@
 template<typename parent_class> class thread_pool {
 public:
     thread_pool() = delete;
-    thread_pool(parent_class *parent, size_t num_workers);
+    thread_pool(parent_class *parent, unsigned num_workers);
     ~thread_pool();
 
-    void execute(void(parent_class::*func)(size_t thread_index));
+    void execute(void(parent_class::*func)(unsigned thread_index));
 
 private:
-    void loop(size_t thread_index);
+    void loop(unsigned thread_index);
 
     parent_class *parent;
     std::vector<std::thread> threads;
@@ -24,8 +24,8 @@ private:
     std::condition_variable threads_activated;
     std::condition_variable threads_completed;
     std::vector<bool> thread_busy;
-    size_t num_threads_busy = 0;
-    void(parent_class::*work_function)(size_t thread_index) = nullptr;
+    unsigned num_threads_busy = 0;
+    void(parent_class::*work_function)(unsigned thread_index) = nullptr;
     bool stopping = false;
 };
